requirements.txt
tqdm
nltk
transformers
torch
numpy
google-cloud-translate
python-multipart
uvicorn
fastapi
soundfile
python-dotenv
assemblyai